refactor: extract common logic to validator class
