+++ src/main/java/com/gotocompany/depot/maxcompute/converter/TimestampProtobufMaxComputeConverter.java
@@ -3,18 +3,22 @@ package com.gotocompany.depot.maxcompute.converter;
 import com.aliyun.odps.type.TypeInfo;
 import com.aliyun.odps.type.TypeInfoFactory;
 import com.google.protobuf.Descriptors;
+import com.google.protobuf.Message;
 import com.gotocompany.depot.config.MaxComputeSinkConfig;
 import com.gotocompany.depot.maxcompute.model.ProtoPayload;
+import com.gotocompany.depot.maxcompute.util.LocalDateTimeValidator;
 
 import java.sql.Timestamp;
-import java.time.LocalDateTime;
 
 public class TimestampProtobufMaxComputeConverter implements ProtobufMaxComputeConverter {
 
-    private final TimestampNTZProtobufMaxComputeConverter timestampNtzProtobufMaxComputeConverter;
+    private static final String SECONDS = "seconds";
+    private static final String NANOS = "nanos";
+
+    private final LocalDateTimeValidator localDateTimeValidator;
 
     public TimestampProtobufMaxComputeConverter(MaxComputeSinkConfig maxComputeSinkConfig) {
-        this.timestampNtzProtobufMaxComputeConverter = new TimestampNTZProtobufMaxComputeConverter(maxComputeSinkConfig);
+        this.localDateTimeValidator = new LocalDateTimeValidator(maxComputeSinkConfig);
     }
 
     @Override
@@ -24,7 +28,10 @@ public class TimestampProtobufMaxComputeConverter implements ProtobufMaxComputeC
 
     @Override
     public Object convertSingularPayload(ProtoPayload protoPayload) {
-        return Timestamp.valueOf((LocalDateTime) timestampNtzProtobufMaxComputeConverter.convertSingularPayload(protoPayload));
+        Message message = (Message) protoPayload.getParsedObject();
+        long seconds = (long) message.getField(message.getDescriptorForType().findFieldByName(SECONDS));
+        int nanos = (int) message.getField(message.getDescriptorForType().findFieldByName(NANOS));
+        return Timestamp.valueOf(localDateTimeValidator.parseAndValidate(seconds, nanos, protoPayload.getFieldDescriptor().getName(), protoPayload.isRootLevel()));
     }
 
 }
